fix: 技術的負債の解消 (Issue #47)
- cast_engagementsテーブルに時間重複禁止制約を追加
- payroll_revenue_factsビューを作成し、給与計算を最適化
- bottle_keep_notificationsの多態参照を正規化
- bottle_keep_locationsの在庫数自動化トリガーを改善
- 重複コンポーネント(CastAssignmentDialog)を削除
- GuestOrderManagementコンポーネントを整理
- バックアップファイル(.bak)を削除
- payroll.serviceを新しいビューを使用するよう修正
