+++ src/app/globals.css
@@ -8,15 +8,7 @@
   --background-end-rgb: 255, 255, 255;
 }
 
-/* ダークモードを無効化
-@media (prefers-color-scheme: dark) {
-  :root {
-    --foreground-rgb: 255, 255, 255;
-    --background-start-rgb: 0, 0, 0;
-    --background-end-rgb: 0, 0, 0;
-  }
-}
-*/
+/*
 
 body {
   color: rgb(var(--foreground-rgb));
