+++ src/services/qr-code/index.ts
@@ -30,60 +30,8 @@ export {
  * 既存のコードが動作するよう、従来のインターフェースを提供
  */
 export class QRCodeService {
-  // 循環インポートを回避するため、一時的に無効化
-  // 全メソッドを一時的に無効化
-  /*
-  // private generationService = qrGenerationService;
-  // private attendanceService = qrAttendanceService;
-  // private statisticsService = qrStatisticsService;
-  // private managementService = qrManagementService;
-
-  // QR生成・検証
-  async generateQRCode(
-    request: import("@/types/qr-code.types").GenerateQRCodeRequest
-  ) {
-    return this.generationService.generateQRCode(request);
-  }
-
-  async validateQRCode(qrData: string) {
-    return this.generationService.validateQRCode(qrData);
-  }
-
-  // 出勤記録
-  async recordAttendance(
-    request: import("@/types/qr-code.types").QRAttendanceRequest
-  ) {
-    return this.attendanceService.recordAttendance(request);
-  }
-
-  async getQRAttendanceHistory(params: {
-    staffId?: string;
-    startDate?: Date;
-    endDate?: Date;
-    limit?: number;
-    offset?: number;
-  }) {
-    return this.attendanceService.getQRAttendanceHistory(params);
-  }
-
-  // 統計・分析
-  async getQRCodeStats() {
-    return this.statisticsService.getQRCodeStats();
-  }
-
-  // 管理・設定
-  async getStaffQRInfo(staffId: string) {
-    return this.managementService.getStaffQRInfo(staffId);
-  }
-
-  async getManagementData() {
-    return this.managementService.getManagementData();
-  }
-
-  async getScanHistory(params: { staffId?: string; limit?: number }) {
-    return this.managementService.getScanHistory(params);
-  }
-  */
+  // 循環インポートを回避するため、現在は空のクラス
+  // 将来的に必要に応じて実装を追加
 }
 
 // 既存コード用のシングルトンインスタンス
